Pascal Hargrove - SMU Bootcamp Homework due 7/29
I create a div about the social media icons and gave it a display:block attribute in CSS thinking that I'd be able to center that div with a simple "margin: 0 auto;" attribute added, but it wouldn't work. Need your help with that because I'm doing something wrong.

Also, I'm having a feeling that I did not set up the Gallery page correctly in regards to the colored labeled divs that are place on top of each image. I used negative margin bottom attributes to put them all in place. I'm having a feeling that's not correct. If you could suggest a better more efficient way to do this, I'd greatly appreciate it!

Also, I'm more much familiar with Bootstrap because I've used it so much in the past few years. I'm not an expert, but it's fun to use. Once of the reason I took this web development bootcamp is because I want to know how to code all the types of things that Bootstrap can do myself. Excited to finally learn JavaScript in a deep dive learning plan.