PascalHargrove_Homework_072917_v2.zip
I added the <body> elements to all 3 of my html files, and I also changed an ID to a class on the gallery page because I had used it more than once and didn't realize it until after I had submitted the first submission yesterday. An updated zip file is attached. Thanks!

Pascal