PascalHargrove_BootstrapHomework_080517.zip
I still have a lot of work to do on this. So I'm not anywhere near done by yet. I got behind this week because of my car wreck this past Tuesday and having to watch/join the class via the live stream on Thursday night. So this Bootstrap version is not quite there yet. I'll post an updated version of it by this Sunday. Thanks!